Changed ~dp0 to %cd% in scripts; variable updates
Added "-major 0 -minor 0 -micro 0" to the "buildCommand" variable which lets you decide what version number the built CIA will have. They're in order, so it goes major.minor.micro.

~dp0 was changed because ~dp0 points at the batch file's directory and not the current running directory of the batch file. Modded ROMs seemed to build fine regardless for an unknown reason (possibly because it ran in the dump folder rather than the modded build folder), but the fix was implemented so files were more predictable. 

Documenting here that "--not-update-exefs-hash --not-update-romfs-hash --not-encrypt" can be removed from the "Build_CIA" script to enable their respective options. I couldn't find a clean way to do this with variables while still retaining documentation of the commands, so you'll just have to edit them in the script itself for now.